refactor(loan): change TotalInterest and MonthlyEMI to decimal for accurate financial calculations
Changed TotalInterest and MonthlyEMI fields from int to decimal in LoanDtos.cs and corresponding methods in LoanService.cs to support fractional values. Updated calculation logic to use RoundToCurrency() for proper rounding and refactored methods to reuse existing calculations, improving precision in loan interest and EMI computations.

+++ MoneyBoard.Application/DTOs/LoanDtos.cs
@@ -48,8 +48,8 @@ namespace MoneyBoard.Application.DTOs
         public LoanStatus Status { get; set; }
         public string? Notes { get; set; }
         public decimal TotalAmount { get; set; }
-        public int TotalInterest { get; set; }
-        public int MonthlyEMI { get; set; }
+        public decimal TotalInterest { get; set; }
+        public decimal MonthlyEMI { get; set; }
 
         // Flag to indicate if loan has started repayments (affects editability)
         public bool HasRepaymentStarted { get; set; }
